extract logic into custom react hook
