+++ ui/src/pages/DownloadPage/DownloadPage.tsx
@@ -1,124 +1,13 @@
-import React, { useState } from 'react';
 import Layout from '../../components/Layout';
-import { generateCsvData } from '../../utils/downloadPageController';
-import { UpdateDocumentResponse } from '../../utils/api';
+import { ExtractedData } from '../../utils/api';
+import { useDownloadPage } from './useDownloadPage.ts';
 
 interface DownloadPageProps {
   signOut: () => Promise<void>;
 }
 
 export default function DownloadPage({ signOut }: DownloadPageProps) {
-  const [verifiedData] = useState<UpdateDocumentResponse>(() => {
-    const storedData = sessionStorage.getItem('verifiedData');
-    return storedData ? JSON.parse(storedData)?.updated_document : {};
-  });
-
-  function displayPreviewTable() {
-    if (!verifiedData.extracted_data) {
-      return <p>No extracted data available</p>;
-    }
-
-    return (
-      <table className="usa-table">
-        <thead>
-          <tr>
-            <th scope="col">Field</th>
-            <th scope="col">Value</th>
-            <th scope="col">Confidence</th>
-          </tr>
-        </thead>
-        <tbody>
-          {Object.entries(verifiedData.extracted_data)
-            .sort(([keyA], [keyB]) =>
-              keyA.localeCompare(keyB, undefined, { numeric: true })
-            )
-            .map(([key, field]) => {
-              return (
-                <tr key={key}>
-                  <td>{key}</td>
-                  <td> {field.value ? field.value : ''}</td>
-                  <td>
-                    {field.confidence
-                      ? parseFloat(field.confidence).toFixed(2) + '%'
-                      : ''}
-                  </td>
-                </tr>
-              );
-            })}
-        </tbody>
-      </table>
-    );
-  }
-
-  function downloadCSV() {
-    if (!verifiedData.extracted_data) {
-      console.error('No data available for CSV download');
-      return;
-    }
-
-    const csvContent = generateCsvData(verifiedData.extracted_data);
-
-    downloadData(csvContent, 'text/csv', 'document.csv');
-  }
-
-  function downloadJSON() {
-    if (!verifiedData.extracted_data) {
-      console.error('No data available for JSON download');
-      return;
-    }
-
-    const jsonContent = JSON.stringify(verifiedData, null, 2);
-
-    downloadData(jsonContent, 'application/json', 'document.json');
-  }
-
-  async function downloadData(
-    content: string,
-    contentType: string,
-    filename: string
-  ) {
-    const blob = new Blob([content], { type: contentType });
-    const url = URL.createObjectURL(blob);
-
-    try {
-      const link = document.createElement('a');
-      link.href = url;
-      link.download = filename;
-      link.click();
-    } finally {
-      setTimeout(() => {
-        URL.revokeObjectURL(url);
-      }, 1000); // Small delay to ensure download starts
-    }
-  }
-
-  function handleDownloadSubmit(event: React.FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-
-    // get the selected file type (CSV or JSON) from the radio buttons
-    const selectedElement = document.querySelector(
-      "input[name='download-file-type']:checked"
-    ) as HTMLInputElement | null;
-
-    if (!selectedElement) {
-      console.error('No file type selected');
-      return;
-    }
-
-    const fileType = selectedElement.value;
-
-    if (!fileType) {
-      console.error('No file type selected');
-      return;
-    }
-
-    // download function based on the selected file type
-    if (fileType === 'csv') {
-      downloadCSV();
-    } else {
-      downloadJSON();
-    }
-  }
+  const { handleDownloadSubmit, verifiedData } = useDownloadPage();
 
   return (
     <Layout signOut={signOut}>
@@ -226,9 +115,46 @@ export default function DownloadPage({ signOut }: DownloadPageProps) {
         <div id="preview-section">
           <h2 id="preview-section-title">File Preview</h2>
           <h3 id="preview-section-file-name">File name</h3>
-          <div>{displayPreviewTable()}</div>
+          <div>{displayPreviewTable(verifiedData.extracted_data)}</div>
         </div>
       </div>
     </Layout>
+  );
+}
+
+function displayPreviewTable(extracted_data?: ExtractedData) {
+  if (!extracted_data) {
+    return <p>No extracted data available</p>;
+  }
+
+  return (
+    <table className="usa-table">
+      <thead>
+        <tr>
+          <th scope="col">Field</th>
+          <th scope="col">Value</th>
+          <th scope="col">Confidence</th>
+        </tr>
+      </thead>
+      <tbody>
+        {Object.entries(extracted_data)
+          .sort(([keyA], [keyB]) =>
+            keyA.localeCompare(keyB, undefined, { numeric: true })
+          )
+          .map(([key, field]) => {
+            return (
+              <tr key={key}>
+                <td>{key}</td>
+                <td> {field.value ? field.value : ''}</td>
+                <td>
+                  {field.confidence
+                    ? parseFloat(field.confidence).toFixed(2) + '%'
+                    : ''}
+                </td>
+              </tr>
+            );
+          })}
+      </tbody>
+    </table>
   );
 }
